# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Go Tasks API
  description: |
    Task and habit management API. Users register and manage their own tasks,
    categories, daily logs, and answers. All data is scoped to the authenticated user.

    ## Authentication

    Protected endpoints require a valid JWT access token in the `Authorization` header
    using the Bearer scheme. Access tokens are RS256-signed and valid for 15 minutes.

    **Token Claims:**
    - `sub` — user ID (UUID)
    - `iss` — issuer (configured via JWT_ISSUER)
    - `aud` — audience (configured via JWT_AUDIENCE)
    - `exp` — expiration time
    - `nbf` — not before time
    - `iat` — issued at time
    - `jti` — unique token ID (UUID)

    ## Refresh Tokens

    Refresh tokens are valid for 1 hour and support rotation. On use, the old token
    is deleted and a new one is issued. Logout revokes the refresh token and adds
    the access token's jti to a blocklist.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/categories:
    get:
      tags:
        - Category
      summary: List categories
      operationId: listCategories
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Category
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Category with this name already exists (case-insensitive)
  /api/v1/categories/{id}:
    get:
      tags:
        - Category
      summary: Get a category
      operationId: getCategory
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Category
      summary: Update a category
      operationId: updateCategory
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Category with this name already exists (case-insensitive)

    delete:
      tags:
        - Category
      summary: Soft delete a category
      description: |
        Sets is_active to false. Returns 409 if already inactive or if the
        category has active tasks (deactivate tasks first).
      operationId: deleteCategory
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Category deactivated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Category is already inactive or has active tasks
  /api/v1/categories/{id}/permanent:
    delete:
      tags:
        - Category
      summary: Permanently delete a category
      description: Hard deletes an inactive category. Returns 409 if still active.
      operationId: permanentDeleteCategory
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Category permanently deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot permanently delete an active category
  /api/v1/categories/{id}/reactivate:
    post:
      tags:
        - Category
      summary: Reactivate a category
      description: Sets is_active to true. Returns 409 if already active or name collision.
      operationId: reactivateCategory
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Category reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Category is already active or another active category has this name
  /api/v1/categories/inactive:
    get:
      tags:
        - Category
      summary: List inactive categories
      operationId: listInactiveCategories
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of inactive categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/categories/bulk-delete:
    post:
      tags:
        - Category
      summary: Bulk soft delete categories
      description: Soft deletes multiple categories. Inactive IDs are ignored.
      operationId: bulkDeleteCategories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          description: Invalid request (empty list, too many IDs, invalid UUID)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/categories/bulk-permanent-delete:
    post:
      tags:
        - Category
      summary: Bulk permanently delete categories
      description: Hard deletes multiple inactive categories. Active IDs are ignored.
      operationId: bulkPermanentDeleteCategories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk permanent delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkPermanentDeleteResponse"
        "400":
          description: Invalid request (empty list, too many IDs, invalid UUID)
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Task endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT access token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
          description: Category name (unique per user among active categories, case-insensitive)
        description:
          type: string
          maxLength: 500
        colour:
          type: string
          pattern: "^#[0-9a-f]{6}$"
          description: Hex colour code (e.g., #ff0000). Stored in lower-case.
          example: "#808080"
        is_active:
          type: boolean
          description: Whether the category is active (false = soft deleted)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Category name (unique per user, case-insensitive, whitespace-trimmed)
        description:
          type: string
          maxLength: 500
        colour:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          description: |
            Hex colour code (e.g., #FF0000 or #ff0000).
            Accepts both upper and lower case on input; stored as lower-case.
            If omitted on create, defaults to #808080.
            If omitted on update, keeps the existing colour.
          example: "#ff0000"

    # Task schemas
...
    BulkDeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
          description: List of IDs to delete (1-100 IDs, duplicates deduplicated server-side)
    BulkDeleteResponse:
      type: object
      properties:
        requested:
          type: integer
          description: Number of unique IDs requested (after deduplication)
        soft_deleted:
          type: integer
          description: Number of items soft deleted (set to inactive)
    BulkPermanentDeleteResponse:
      type: object
      properties:
        requested:
          type: integer
          description: Number of unique IDs requested (after deduplication)
        permanently_deleted:
          type: integer
          description: Number of items permanently deleted
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Resource not found